+++ powersched/compile.sh
@@ -2,6 +2,8 @@
 
 set -e
 
+[ -f all/usr/sbin/powersched ] && exit 0
+
 mkdir -p all/usr/sbin
 make -C src clean all
 cp src/powersched all/usr/sbin
